x64 bug fix
git-svn-id: https://svn.java.net/svn/com4j~svn/trunk/com4j@390 590aa308-3dd6-4388-9d27-cebafe3e2bdc

+++ runtime/src/com4j/NativeType.java
@@ -280,7 +280,7 @@ public enum NativeType {
             // And then, after the call, this method (toJava) gets called from com4j.StandardComMethod.invoke(int, Object[])
             // So we create a new Wrapper object for an interface pointer. But we can not assume, that the native code of the third party
             // software called addRef for a pointer we passed in! So we call addRef ourself. (see issues 25 and 36)
-            Native.addRef((Integer) param);
+            Native.addRef((Long) param);
 
             return Wrapper.create( (Class<? extends Com4jObject>)type, (Long)param );
         }
